third_party/WebKit/LayoutTests/battery-status/detached-no-crash.html

<!DOCTYPE HTML>
<body>
<head>
<script src="../resources/js-test.js"></script>
</head>
<script>
description("Accessing navigator.getBattery() on a closed window.");

window.jsTestIsAsync = true;

var w;
var nav;

function processMessage(event) {
    if (event.data == "opened") {
        nav = w.navigator;
        shouldBeTrue("typeof(nav.getBattery()) == 'object'");
        w.close();
        w = null;
    } else if (event.data == "closed") {
        shouldBeUndefined("nav.getBattery()");
        finishJSTest();
    }
}

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
}
w = window.open('../resources/window-postmessage-open-close.html');
window.addEventListener("message", processMessage, false);
</script>
</body>
